+++ worker/include/RTC/DataConsumer.hpp
@@ -7,6 +7,7 @@
 #include "PayloadChannel/PayloadChannelRequest.hpp"
 #include "PayloadChannel/PayloadChannelSocket.hpp"
 #include "RTC/SctpDictionaries.hpp"
+#include "RTC/Shared.hpp"
 #include <nlohmann/json.hpp>
 #include <string>
 
@@ -47,6 +48,7 @@ namespace RTC
 
 	public:
 		DataConsumer(
+		  RTC::Shared* shared,
 		  const std::string& id,
 		  const std::string& dataProducerId,
 		  RTC::SctpAssociation* sctpAssociation,
@@ -81,6 +83,7 @@ namespace RTC
 		void SctpAssociationConnected();
 		void SctpAssociationClosed();
 		void SctpAssociationBufferedAmount(uint32_t bufferedAmount);
+		void SctpAssociationSendBufferFull();
 		void DataProducerClosed();
 		void SendMessage(uint32_t ppid, const uint8_t* msg, size_t len, onQueuedCallback* = nullptr);
 
@@ -99,6 +102,7 @@ namespace RTC
 
 	private:
 		// Passed by argument.
+		RTC::Shared* shared{ nullptr };
 		RTC::SctpAssociation* sctpAssociation{ nullptr };
 		RTC::DataConsumer::Listener* listener{ nullptr };
 		size_t maxMessageSize{ 0u };