+++ worker/include/RTC/DataProducer.hpp
@@ -7,6 +7,7 @@
 #include "PayloadChannel/PayloadChannelSocket.hpp"
 #include "RTC/RTCP/Packet.hpp"
 #include "RTC/SctpDictionaries.hpp"
+#include "RTC/Shared.hpp"
 #include <nlohmann/json.hpp>
 #include <string>
 
@@ -36,7 +37,11 @@ namespace RTC
 
 	public:
 		DataProducer(
-		  const std::string& id, size_t maxMessageSize, RTC::DataProducer::Listener* listener, json& data);
+		  RTC::Shared* shared,
+		  const std::string& id,
+		  size_t maxMessageSize,
+		  RTC::DataProducer::Listener* listener,
+		  json& data);
 		virtual ~DataProducer();
 
 	public:
@@ -66,6 +71,7 @@ namespace RTC
 
 	private:
 		// Passed by argument.
+		RTC::Shared* shared{ nullptr };
 		size_t maxMessageSize{ 0u };
 		RTC::DataProducer::Listener* listener{ nullptr };
 		// Others.