+++ worker/src/RTC/WebRtcTransport.cpp
@@ -2,11 +2,9 @@
 // #define MS_LOG_DEV_LEVEL 3
 
 #include "RTC/WebRtcTransport.hpp"
-#include "ChannelMessageHandlers.hpp"
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
 #include "Utils.hpp"
-#include "Channel/ChannelNotifier.hpp"
 #include <cmath> // std::pow()
 
 namespace RTC
@@ -29,8 +27,9 @@ namespace RTC
 
 	/* Instance methods. */
 
-	WebRtcTransport::WebRtcTransport(const std::string& id, RTC::Transport::Listener* listener, json& data)
-	  : RTC::Transport::Transport(id, listener, data)
+	WebRtcTransport::WebRtcTransport(
+	  RTC::Shared* shared, const std::string& id, RTC::Transport::Listener* listener, json& data)
+	  : RTC::Transport::Transport(shared, id, listener, data)
 	{
 		MS_TRACE();
 
@@ -206,7 +205,7 @@ namespace RTC
 			this->dtlsTransport = new RTC::DtlsTransport(this);
 
 			// NOTE: This may throw.
-			ChannelMessageHandlers::RegisterHandler(
+			this->shared->channelMessageRegistrator->RegisterHandler(
 			  this->id,
 			  /*channelRequestHandler*/ this,
 			  /*payloadChannelRequestHandler*/ this,
@@ -248,12 +247,13 @@ namespace RTC
 	 * This constructor is used when the WebRtcTransport uses a WebRtcServer.
 	 */
 	WebRtcTransport::WebRtcTransport(
+	  RTC::Shared* shared,
 	  const std::string& id,
 	  RTC::Transport::Listener* listener,
 	  WebRtcTransportListener* webRtcTransportListener,
 	  std::vector<RTC::IceCandidate>& iceCandidates,
 	  json& data)
-	  : RTC::Transport::Transport(id, listener, data),
+	  : RTC::Transport::Transport(shared, id, listener, data),
 	    webRtcTransportListener(webRtcTransportListener), iceCandidates(iceCandidates)
 	{
 		MS_TRACE();
@@ -274,7 +274,7 @@ namespace RTC
 			this->webRtcTransportListener->OnWebRtcTransportCreated(this);
 
 			// NOTE: This may throw.
-			ChannelMessageHandlers::RegisterHandler(
+			this->shared->channelMessageRegistrator->RegisterHandler(
 			  this->id,
 			  /*channelRequestHandler*/ this,
 			  /*payloadChannelRequestHandler*/ this,
@@ -300,7 +300,7 @@ namespace RTC
 	{
 		MS_TRACE();
 
-		ChannelMessageHandlers::UnregisterHandler(this->id);
+		this->shared->channelMessageRegistrator->UnregisterHandler(this->id);
 
 		// Must delete the DTLS transport first since it will generate a DTLS alert
 		// to be sent.
@@ -1304,7 +1304,7 @@ namespace RTC
 
 		this->iceServer->GetSelectedTuple()->FillJson(data["iceSelectedTuple"]);
 
-		Channel::ChannelNotifier::Emit(this->id, "iceselectedtuplechange", data);
+		this->shared->channelNotifier->Emit(this->id, "iceselectedtuplechange", data);
 	}
 
 	inline void WebRtcTransport::OnIceServerConnected(const RTC::IceServer* /*iceServer*/)
@@ -1318,7 +1318,7 @@ namespace RTC
 
 		data["iceState"] = "connected";
 
-		Channel::ChannelNotifier::Emit(this->id, "icestatechange", data);
+		this->shared->channelNotifier->Emit(this->id, "icestatechange", data);
 
 		// If ready, run the DTLS handler.
 		MayRunDtlsTransport();
@@ -1341,7 +1341,7 @@ namespace RTC
 
 		data["iceState"] = "completed";
 
-		Channel::ChannelNotifier::Emit(this->id, "icestatechange", data);
+		this->shared->channelNotifier->Emit(this->id, "icestatechange", data);
 
 		// If ready, run the DTLS handler.
 		MayRunDtlsTransport();
@@ -1364,7 +1364,7 @@ namespace RTC
 
 		data["iceState"] = "disconnected";
 
-		Channel::ChannelNotifier::Emit(this->id, "icestatechange", data);
+		this->shared->channelNotifier->Emit(this->id, "icestatechange", data);
 
 		// If DTLS was already connected, notify the parent class.
 		if (this->dtlsTransport->GetState() == RTC::DtlsTransport::DtlsState::CONNECTED)
@@ -1384,7 +1384,7 @@ namespace RTC
 
 		data["dtlsState"] = "connecting";
 
-		Channel::ChannelNotifier::Emit(this->id, "dtlsstatechange", data);
+		this->shared->channelNotifier->Emit(this->id, "dtlsstatechange", data);
 	}
 
 	inline void WebRtcTransport::OnDtlsTransportConnected(
@@ -1428,7 +1428,7 @@ namespace RTC
 			data["dtlsState"]      = "connected";
 			data["dtlsRemoteCert"] = remoteCert;
 
-			Channel::ChannelNotifier::Emit(this->id, "dtlsstatechange", data);
+			this->shared->channelNotifier->Emit(this->id, "dtlsstatechange", data);
 
 			// Tell the parent class.
 			RTC::Transport::Connected();
@@ -1453,7 +1453,7 @@ namespace RTC
 
 		data["dtlsState"] = "failed";
 
-		Channel::ChannelNotifier::Emit(this->id, "dtlsstatechange", data);
+		this->shared->channelNotifier->Emit(this->id, "dtlsstatechange", data);
 	}
 
 	inline void WebRtcTransport::OnDtlsTransportClosed(const RTC::DtlsTransport* /*dtlsTransport*/)
@@ -1467,7 +1467,7 @@ namespace RTC
 
 		data["dtlsState"] = "closed";
 
-		Channel::ChannelNotifier::Emit(this->id, "dtlsstatechange", data);
+		this->shared->channelNotifier->Emit(this->id, "dtlsstatechange", data);
 
 		// Tell the parent class.
 		RTC::Transport::Disconnected();